requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-chroma
chromadb
python-dotenv
markdown